Rework Chart AppVersion and Image tag
This commit tries to rework how the image tags are managed to simplify
the release process

1) Use of "latest" tag is not recommended.
2) To make changing chart version easier for every subsequent
release as this involves making a change in a single place i.e.
changing AppVersion in Charts.yaml & the value for image tag can be
independent of chart's AppVersion if the tag value is set in
values.yaml.