Initial values.yaml for GKE
Initial values.yaml for Google K*8 Engine
This is copy of the values files from neuvector-helm-eks payg-llc branch
with the following changes:
change registry: to something that is goolgle. Not sure I have the correct
setting, the docs are not clear here.

change neuvector-csp-billing-adapter-llc to neuvector-csp-billing-adapter
since google will have only a single billing authority.