Switch to using images tags instead of digests for azure
The tool that Microsoft uses to create the CNAB bundle for the Azure Marketplace offer has been enhanced to support image tags as well as digests.

Azure-Samples/kubernetes-offer-samples#31

This PR switchws the azure image handling to tags to be consistent with the other platforms.

Additionally, since the scanner is not installed as part of the Azure Marketplace offer, but added after, it was removed from the Azure section.

https://suse-enceladus.github.io/marketplace-docs/neuvector-prime/azure/